refactor: relocate messages.ts to src/agent/runners/pi/ (#632 segment 1)
Pure relocate, no logic change. messages.ts constructs AgentMessage from
@mariozechner/pi-agent-core — pi-bound by type, so it lives under the pi
runner directory rather than at agent/ top-level.

This establishes the convention that runner-specific code lives under
src/agent/runners/<runner>/. Cross-runner code (events, manager, init,
run-adapter, workspace, ...) stays at agent/ top-level.

- src/legacy/core/messages.{ts,test.ts} → src/agent/runners/pi/
- 7 importers updated (2 external in plugins/, 5 internal in core/)

Acknowledged leak: discord.ts and http/sessions.ts now import from
agent/runners/pi/ to construct user input messages, because pi's
AgentMessage is currently the universal internal format. A future PR
will introduce an isotopes-internal Message type with per-runner
adapters; until then, the import path makes the leak visible.

claude.ts has a duplicate buildAssistantMessage inline; dedup deferred
to its own PR.

tsc clean; 1179/1179 tests pass.
